+++ test/blackbox-tests/test-cases/github734/run.t
@@ -2,9 +2,6 @@
   File "src/dune", line 4, characters 10-17:
   4 |  (c_names stubs/x))
                 ^^^^^^^
-  Warning: relative part of stub are no longer necessary and are ignored.
-  File "src/dune", line 4, characters 10-17:
-  4 |  (c_names stubs/x))
-                ^^^^^^^
-  Error: x does not exist as a C source. x.c must be present
+  Error: relative part of stub is not necessary and should be removed. To
+  include sources in subdirectories, use the include_subdirs stanza
   [1]