+++ CHANGES.md
@@ -27,13 +27,16 @@
   would be allowed for unwrapped libraries (#2442, @rgrinberg)
 
 - mli only modules must now be explicitly declared. This was previously a
-  warning that is now an error. (#2442, @rgrinberg)
+  warning and is now an error. (#2442, @rgrinberg)
 
 - Modules filtered out from the module list via the Ordered Set Language must
   now be actual modules. (#2442, @rgrinberg)
 
 - Actions which introduce targets where new targets are forbidden (e.g.
   preprocessing) are now an error instead of a warning. (#2442, @rgrinberg)
+
+- Stub names are no longer allowed relative paths. This was previously a warning
+  and is now an error (#2443, @rgrinberg).
 
 1.11.0 (23/07/2019)
 -------------------